Flatten and document shared model
