+++ gnd/src/main/java/com/google/android/gnd/persistence/local/room/FeatureEntity.java
@@ -22,8 +22,8 @@ import androidx.room.Embedded;
 import androidx.room.Entity;
 import androidx.room.Index;
 import androidx.room.PrimaryKey;
-import com.google.android.gnd.persistence.shared.FeatureMutation;
 import com.google.android.gnd.vo.Feature;
+import com.google.android.gnd.vo.FeatureMutation;
 import com.google.android.gnd.vo.Project;
 import com.google.auto.value.AutoValue;
 import com.google.auto.value.AutoValue.CopyAnnotations;
