+++ gnd/src/main/java/com/google/android/gnd/persistence/local/room/LocalDataStoreException.java
@@ -16,7 +16,7 @@
 
 package com.google.android.gnd.persistence.local.room;
 
-import com.google.android.gnd.persistence.shared.Mutation;
+import com.google.android.gnd.vo.Mutation;
 
 /**
  * Checked exception used internally to ensure we catch and properly handle errors in the Room local
