+++ gnd/src/main/java/com/google/android/gnd/persistence/local/room/RecordEntity.java
@@ -25,9 +25,9 @@ import androidx.room.Entity;
 import androidx.room.ForeignKey;
 import androidx.room.Index;
 import androidx.room.PrimaryKey;
-import com.google.android.gnd.persistence.shared.RecordMutation;
 import com.google.android.gnd.vo.Feature;
 import com.google.android.gnd.vo.Record;
+import com.google.android.gnd.vo.RecordMutation;
 import com.google.android.gnd.vo.ResponseMap;
 import com.google.auto.value.AutoValue;
 import com.google.auto.value.AutoValue.CopyAnnotations;
