+++ gnd/src/main/java/com/google/android/gnd/persistence/local/room/RecordMutationEntity.java
@@ -25,17 +25,14 @@ import androidx.room.Entity;
 import androidx.room.ForeignKey;
 import androidx.room.Index;
 import androidx.room.PrimaryKey;
-import com.google.android.gnd.persistence.shared.RecordMutation;
-import com.google.android.gnd.persistence.shared.ResponseDelta;
+import com.google.android.gnd.vo.RecordMutation;
+import com.google.android.gnd.vo.ResponseDelta;
 import com.google.auto.value.AutoValue;
 import com.google.auto.value.AutoValue.CopyAnnotations;
 import com.google.common.collect.ImmutableList;
 import org.json.JSONObject;
 
-/**
- * Representation of a {@link com.google.android.gnd.persistence.shared.RecordMutation} in local
- * data store.
- */
+/** Representation of a {@link RecordMutation} in local data store. */
 @AutoValue
 @Entity(
     tableName = "record_mutation",
