+++ gnd/src/main/java/com/google/android/gnd/persistence/remote/RemoteDataStore.java
@@ -16,10 +16,10 @@
 
 package com.google.android.gnd.persistence.remote;
 
-import com.google.android.gnd.persistence.shared.Mutation;
 import com.google.android.gnd.system.AuthenticationManager.User;
 import com.google.android.gnd.vo.Feature;
 import com.google.android.gnd.vo.FeatureUpdate.RecordUpdate.ResponseUpdate;
+import com.google.android.gnd.vo.Mutation;
 import com.google.android.gnd.vo.Project;
 import com.google.android.gnd.vo.Record;
 import com.google.common.collect.ImmutableCollection;
