+++ gnd/src/main/java/com/google/android/gnd/ui/common/BindingAdapters.java
@@ -16,20 +16,20 @@
 
 package com.google.android.gnd.ui.common;
 
-import androidx.databinding.BindingAdapter;
-import androidx.databinding.DataBindingUtil;
-import androidx.databinding.ViewDataBinding;
-import androidx.annotation.Nullable;
-import com.google.android.gnd.vo.Record.Response;
-import com.google.android.material.textfield.TextInputEditText;
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.view.View;
+import androidx.annotation.Nullable;
+import androidx.databinding.BindingAdapter;
+import androidx.databinding.DataBindingUtil;
+import androidx.databinding.ViewDataBinding;
 import com.google.android.gms.common.SignInButton;
 import com.google.android.gnd.databinding.MultipleChoiceInputFieldBinding;
 import com.google.android.gnd.databinding.TextInputFieldBinding;
 import com.google.android.gnd.ui.editrecord.MultipleChoiceFieldLayout;
-import com.google.android.gnd.vo.Form.Field;
+import com.google.android.gnd.vo.Field;
+import com.google.android.gnd.vo.Response;
+import com.google.android.material.textfield.TextInputEditText;
 import java8.util.function.Consumer;
 
 public class BindingAdapters {
