+++ gnd/src/main/java/com/google/android/gnd/ui/editrecord/EditRecordViewModel.java
@@ -29,20 +29,20 @@ import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import com.google.android.gnd.GndApplication;
 import com.google.android.gnd.R;
-import com.google.android.gnd.persistence.shared.Mutation;
-import com.google.android.gnd.persistence.shared.RecordMutation;
-import com.google.android.gnd.persistence.shared.ResponseDelta;
 import com.google.android.gnd.repository.DataRepository;
 import com.google.android.gnd.repository.Persistable;
 import com.google.android.gnd.system.AuthenticationManager;
 import com.google.android.gnd.ui.common.AbstractViewModel;
 import com.google.android.gnd.ui.common.SingleLiveEvent;
-import com.google.android.gnd.vo.Form.Element;
-import com.google.android.gnd.vo.Form.Element.Type;
-import com.google.android.gnd.vo.Form.Field;
+import com.google.android.gnd.vo.Element;
+import com.google.android.gnd.vo.Element.Type;
+import com.google.android.gnd.vo.Field;
+import com.google.android.gnd.vo.Mutation;
 import com.google.android.gnd.vo.Record;
-import com.google.android.gnd.vo.Record.Response;
-import com.google.android.gnd.vo.Record.TextResponse;
+import com.google.android.gnd.vo.RecordMutation;
+import com.google.android.gnd.vo.Response;
+import com.google.android.gnd.vo.ResponseDelta;
+import com.google.android.gnd.vo.TextResponse;
 import com.google.common.collect.ImmutableList;
 import io.reactivex.Completable;
 import io.reactivex.Observable;
