+++ gnd/src/main/java/com/google/android/gnd/ui/editrecord/MultiSelectDialogFactory.java
@@ -20,11 +20,11 @@ import static java8.util.stream.StreamSupport.stream;
 import android.content.Context;
 import androidx.appcompat.app.AlertDialog;
 import com.google.android.gnd.R;
-import com.google.android.gnd.vo.Form.Field;
-import com.google.android.gnd.vo.Form.MultipleChoice;
-import com.google.android.gnd.vo.Form.MultipleChoice.Option;
-import com.google.android.gnd.vo.Record.MultipleChoiceResponse;
-import com.google.android.gnd.vo.Record.Response;
+import com.google.android.gnd.vo.Field;
+import com.google.android.gnd.vo.MultipleChoice;
+import com.google.android.gnd.vo.MultipleChoiceResponse;
+import com.google.android.gnd.vo.Option;
+import com.google.android.gnd.vo.Response;
 import com.google.common.collect.ImmutableList;
 import java.util.List;
 import java8.util.Optional;
