+++ gnd/src/main/java/com/google/android/gnd/ui/home/featuresheet/RecordListItemViewHolder.java
@@ -31,11 +31,11 @@ import butterknife.BindView;
 import butterknife.ButterKnife;
 import com.google.android.gnd.R;
 import com.google.android.gnd.system.AuthenticationManager;
+import com.google.android.gnd.vo.Element;
+import com.google.android.gnd.vo.Field;
 import com.google.android.gnd.vo.Form;
-import com.google.android.gnd.vo.Form.Element;
-import com.google.android.gnd.vo.Form.Field;
 import com.google.android.gnd.vo.Record;
-import com.google.android.gnd.vo.Record.Response;
+import com.google.android.gnd.vo.Response;
 import java.text.DateFormat;
 import java.util.Date;
 import java8.util.Optional;
