+++ gnd/src/main/java/com/google/android/gnd/vo/Element.java
@@ -1,0 +1,48 @@
+/*
+ * Copyright 2019 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.android.gnd.vo;
+
+import com.google.auto.value.AutoOneOf;
+import javax.annotation.Nullable;
+
+/**
+ * Represents an element on a form. The only valid type of form element is currently {@code FIELD};
+ * this abstraction exists for future use (e.g., to support headings, subforms).
+ */
+@AutoOneOf(Element.Type.class)
+public abstract class Element {
+  public enum Type {
+    UNKNOWN,
+    FIELD
+  }
+
+  public abstract Type getType();
+
+  @Nullable
+  public abstract Field getField();
+
+  @Nullable
+  public abstract Object getUnknown();
+
+  public static Element ofField(Field field) {
+    return AutoOneOf_Form_Element.field(field);
+  }
+
+  public static Element ofUnknown() {
+    return AutoOneOf_Form_Element.unknown(new Object());
+  }
+}
