+++ gnd/src/main/java/com/google/android/gnd/vo/Record.java
@@ -16,17 +16,11 @@
 
 package com.google.android.gnd.vo;
 
-import static java8.util.stream.StreamSupport.stream;
-
 import androidx.annotation.Nullable;
 import com.google.android.gnd.system.AuthenticationManager.User;
-import com.google.android.gnd.vo.Form.Field;
-import com.google.android.gnd.vo.Form.MultipleChoice.Option;
 import com.google.auto.value.AutoValue;
-import java.util.List;
-import java8.util.Optional;
-import java8.util.stream.Collectors;
 
+/** Represents a single instance of data collected about a specific {@link Feature}. */
 @AutoValue
 public abstract class Record {
   @Nullable
@@ -82,135 +76,5 @@ public abstract class Record {
     public abstract Builder setResponses(ResponseMap responses);
 
     public abstract Record build();
-  }
-
-  public interface Response {
-
-    String getSummaryText(Field field);
-
-    String getDetailsText(Field field);
-
-    static String toString(Optional<Response> response) {
-      return response.map(Response::toString).orElse("");
-    }
-
-    boolean isEmpty();
-  }
-
-  public static class TextResponse implements Response {
-
-    private String text;
-
-    public TextResponse(String text) {
-      this.text = text;
-    }
-
-    public String getText() {
-      return text;
-    }
-
-    @Override
-    public String getSummaryText(Field field) {
-      return text;
-    }
-
-    @Override
-    public String getDetailsText(Field field) {
-      return text;
-    }
-
-    @Override
-    public boolean isEmpty() {
-      return text.trim().isEmpty();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      if (obj == null || !(obj instanceof TextResponse)) {
-        return false;
-      }
-      return text.equals(((TextResponse) obj).text);
-    }
-
-    @Override
-    public int hashCode() {
-      return text.hashCode();
-    }
-
-    @Override
-    public String toString() {
-      return text;
-    }
-
-    public static Optional<Response> fromString(String text) {
-      return text.isEmpty() ? Optional.empty() : Optional.of(new TextResponse(text));
-    }
-  }
-
-  public static class MultipleChoiceResponse implements Response {
-
-    private List<String> choices;
-
-    public MultipleChoiceResponse(List<String> choices) {
-      this.choices = choices;
-    }
-
-    public List<String> getChoices() {
-      return choices;
-    }
-
-    public Optional<String> getFirstCode() {
-      return stream(choices).findFirst();
-    }
-
-    public boolean isSelected(Option option) {
-      return choices.contains(option.getCode());
-    }
-
-    public String getSummaryText(Field field) {
-      return toString();
-    }
-
-    // TODO: Make these inner classes non-static and access Form directly.
-    public String getDetailsText(Field field) {
-      return stream(choices)
-          .map(code -> field.getMultipleChoice().getOption(code))
-          .filter(Optional::isPresent)
-          .map(Optional::get)
-          .map(Option::getLabel)
-          .sorted()
-          .collect(Collectors.joining(", "));
-    }
-
-    @Override
-    public boolean isEmpty() {
-      return choices.isEmpty();
-    }
-
-    @Override
-    public boolean equals(Object obj) {
-      if (obj == null || !(obj instanceof MultipleChoiceResponse)) {
-        return false;
-      }
-      return choices.equals(((MultipleChoiceResponse) obj).choices);
-    }
-
-    @Override
-    public int hashCode() {
-      return choices.hashCode();
-    }
-
-    @Override
-    public String toString() {
-      return stream(choices).sorted().collect(Collectors.joining(","));
-    }
-
-    public static Optional<Response> fromList(List<String> codes) {
-      if (codes.isEmpty()) {
-        return Optional.empty();
-      } else {
-        return Optional.of(new MultipleChoiceResponse(codes));
-      }
-    }
   }
 }
