+++ gnd/src/main/res/layout/multiple_choice_input_field.xml
@@ -14,75 +14,75 @@
   ~ limitations under the License.
   -->
 <layout
-  xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools">
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools">
   <data>
     <variable
-      name="fragment"
-      type="com.google.android.gnd.ui.editrecord.EditRecordFragment"/>
+        name="fragment"
+        type="com.google.android.gnd.ui.editrecord.EditRecordFragment"/>
     <variable
-      name="field"
-      type="com.google.android.gnd.vo.Form.Field"/>
+        name="field"
+        type="com.google.android.gnd.vo.Field"/>
     <variable
-      name="viewModel"
-      type="com.google.android.gnd.ui.editrecord.EditRecordViewModel"/>
+        name="viewModel"
+        type="com.google.android.gnd.ui.editrecord.EditRecordViewModel"/>
   </data>
 
   <com.google.android.gnd.ui.editrecord.MultipleChoiceFieldLayout
-    android:layout_width="match_parent"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="@dimen/field_vertical_margin"
-    android:layout_marginBottom="@dimen/field_vertical_margin"
-    android:layout_marginStart="@dimen/field_horizontal_margin"
-    android:layout_marginEnd="@dimen/field_horizontal_margin"
-    android:layout_marginLeft="@dimen/field_horizontal_margin"
-    android:layout_marginRight="@dimen/field_horizontal_margin"
-    app:onShowDialog="@{() -> fragment.onShowDialog(field)}">
-
-    <com.google.android.material.textfield.TextInputLayout
-      android:id="@+id/multiple_choice_input_layout"
-      style="@style/EditRecord.Field.Text"
       android:layout_width="match_parent"
       android:layout_height="wrap_content"
-      android:hint="@{field.label}">
+      android:layout_marginTop="@dimen/field_vertical_margin"
+      android:layout_marginBottom="@dimen/field_vertical_margin"
+      android:layout_marginStart="@dimen/field_horizontal_margin"
+      android:layout_marginEnd="@dimen/field_horizontal_margin"
+      android:layout_marginLeft="@dimen/field_horizontal_margin"
+      android:layout_marginRight="@dimen/field_horizontal_margin"
+      app:onShowDialog="@{() -> fragment.onShowDialog(field)}">
+
+    <com.google.android.material.textfield.TextInputLayout
+        android:id="@+id/multiple_choice_input_layout"
+        style="@style/EditRecord.Field.Text"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:hint="@{field.label}">
 
       <!--
         Using deprecated editable="false" as a hack to reuse TextInputEditText styles for multi-
         select box.
        -->
       <com.google.android.material.textfield.TextInputEditText
-        android:id="@+id/multiple_choice_input_edit_text"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:editable="false"
-        android:focusable="true"
-        android:focusableInTouchMode="true"
-        android:text="@{viewModel.responses[field.id]}"
-        android:textColor="@color/field_value"
-        app:errorText="@{viewModel.errors[field.id]}"
-        tools:text="One, two, three, four, five"/>
+          android:id="@+id/multiple_choice_input_edit_text"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          android:editable="false"
+          android:focusable="true"
+          android:focusableInTouchMode="true"
+          android:text="@{viewModel.responses[field.id]}"
+          android:textColor="@color/field_value"
+          app:errorText="@{viewModel.errors[field.id]}"
+          tools:text="One, two, three, four, five"/>
 
     </com.google.android.material.textfield.TextInputLayout>
 
     <ImageButton
-      android:id="@+id/multiple_choice_dropdown_btn"
-      android:layout_width="32dp"
-      android:layout_height="32dp"
-      android:layout_marginRight="4dp"
-      android:layout_gravity="center_vertical|right"
-      android:background="@android:color/transparent"
-      android:elevation="0dp"
-      android:focusable="false"
-      android:focusableInTouchMode="false"
-      android:src="@drawable/ic_arrow_drop_down"/>
+        android:id="@+id/multiple_choice_dropdown_btn"
+        android:layout_width="32dp"
+        android:layout_height="32dp"
+        android:layout_marginRight="4dp"
+        android:layout_gravity="center_vertical|right"
+        android:background="@android:color/transparent"
+        android:elevation="0dp"
+        android:focusable="false"
+        android:focusableInTouchMode="false"
+        android:src="@drawable/ic_arrow_drop_down"/>
 
     <View
-      android:id="@+id/multiple_choice_tap_target"
-      android:layout_width="match_parent"
-      android:layout_height="match_parent"
-      android:alpha="0"
-      android:clickable="true"/>
+        android:id="@+id/multiple_choice_tap_target"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:alpha="0"
+        android:clickable="true"/>
 
   </com.google.android.gnd.ui.editrecord.MultipleChoiceFieldLayout>
 </layout>
