+++ gnd/src/main/res/layout/text_input_field.xml
@@ -21,7 +21,7 @@
   <data>
     <variable
         name="field"
-        type="com.google.android.gnd.vo.Form.Field"/>
+        type="com.google.android.gnd.vo.Field"/>
     <variable
         name="viewModel"
         type="com.google.android.gnd.ui.editrecord.EditRecordViewModel"/>
